requirements.txt
flask
flask-cors
requests
numpy
Pillow
opencv-python-headless
tensorflow-cpu
torch
torchvision
gunicorn